+++ app/src/main/res/layout/listview_layout.xml
@@ -10,25 +10,27 @@
         android:textSize="30sp"
         android:textColor="#000000"
         android:background="#00ff0000"
-        android:layout_gravity="center_vertical"
+        android:gravity="center_vertical"
         android:layout_weight="4"
         />
+
     <ImageButton
-        android:layout_width="60dp"
-        android:layout_height="match_parent"
-        android:src="@drawable/minus"
-        android:scaleType="fitCenter"
+        android:id="@+id/btn_delete"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_gravity="center_vertical"
+        android:layout_weight="1"
         android:adjustViewBounds="true"
         android:background="#00ff0000"
-        android:layout_weight="1"
-        android:id="@+id/btn_delete"
-        />
-        />
+        android:scaleType="fitCenter"
+        android:src="@drawable/minus" />
+
     <ImageButton
-        android:layout_width="60dp"
-        android:layout_height="match_parent"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
         android:id="@+id/btn_info"
         android:src="@drawable/info"
+        android:layout_gravity="center_vertical"
         android:scaleType="fitCenter"
         android:adjustViewBounds="true"
         android:layout_weight="1"
